+++ RunTime/Toolkits/Core/SingletonKit/Singletons/PersistentMonoSingleton.cs
@@ -13,42 +13,19 @@ namespace Framework3.Toolkits.SingletonKit
     /// <summary>
     /// 当场景里包含两个 PersistentMonoSingleton，保留先创建的
     /// </summary>
-    public abstract class PersistentMonoSingleton<TSingleton> : MonoBehaviour where TSingleton : Component
+    public abstract class PersistentMonoSingleton<TMonoSingleton> : MonoSingleton<TMonoSingleton> where TMonoSingleton : PersistentMonoSingleton<TMonoSingleton>
     {
-        protected static TSingleton s_instance;
-
-        public static TSingleton Instance
-        {
-            get
-            {
-                if (s_instance)
-                {
-                    return s_instance;
-                }
-                
-                s_instance = FindFirstObjectByType<TSingleton>();
-                if (s_instance)
-                {
-                    return s_instance;
-                }
-                
-                var obj = new GameObject();
-                s_instance = obj.AddComponent<TSingleton>();
-                return s_instance;
-            }
-        }
-
         // 当场景里包含两个 PersistentMonoSingleton，保留先创建的
-        protected virtual void Awake()
+        public virtual void Awake()
         {
             if (!Application.isPlaying)
             {
                 return;
             }
 
-            if (s_instance == null)
+            if (!s_instance)
             {
-                s_instance = this as TSingleton;
+                s_instance = this as TMonoSingleton;
                 DontDestroyOnLoad(transform.gameObject);
             }
             else