+++ Samples/Toolkits/Core/SingletonKit/8.ScriptableSingletonPropertyExample/MyScriptableA.cs
@@ -9,14 +9,10 @@
 namespace Framework3.Toolkits.SingletonKit.Example._8.ScriptableSingletonPropertyExample
 {
     using UnityEngine;
-    using Framework3.Core;
 
     [CreateAssetMenu(fileName = nameof(MyScriptableA))]
     public class MyScriptableA : ScriptableObject
     {
         public string ScriptableKey;
-        
-        public static MyScriptableA Instance => ScriptableSingletonProperty<MyScriptableA>.InstanceWithLoader(
-            scriptableName => Resources.Load<MyScriptableA>("Scriptable/" + scriptableName));
     }
 }